Organize the sync code
